+++ Web/Design a Library Management System/Program.cs
@@ -1,3 +1,4 @@
+using Design_a_Library_Management_System;
 using DesignLibraryManagementSystem.Models;
 using Library;
 
@@ -8,6 +9,7 @@ builder.Configuration.AddJsonFile("appsettings.json");
 string connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
  builder.Services.AddSingleton(new LibraryContext(connectionString));
 builder.Services.AddControllers();
+builder.Services.AddScoped<SearchService>();
 
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
